backend/requirements.txt
asgiref==3.8.1
betterpath==0.2.2
bidict==0.23.1
click==8.1.7
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
l==0.11.0
packaging==24.1
PyJWT==2.9.0
python-engineio==4.9.1
python-socketio==5.11.3
redis==5.0.8
setuptools==74.1.0
simple-websocket==1.0.0
sqlparse==0.5.1
uvicorn==0.30.6
vcversioner==2.16.0.0
whitenoise==6.7.0
wsproto==1.2.0
zope.interface==7.0.3
